requirements.txt
langchain-google-genai
pillow
streamlit
python-dotenv
langchain
langchain-openai
langchain_experimental
tiktoken
langchain-ollama
langchain_community
langchain-google-genai
sentence-transformers
langchain_chroma
langchainhub
pypdf
rapidocr-onnxruntime